Subject heading override (#18)
* Removed block of code that assigned all locations to "Online" formats

* Update to Bundler version 2 (#8)

Our chruby role installs bundler automatically, but
it gets the newest version (now 2.x). This needs to be updated
in the lock file so that it will work with bundler v2.x.

* Add correct time zone to jdbc connection string

* Allow for subject override using a translation map

This changes the indexing for subject facets so that
they can be translated. The non-translated terms
are still indexed so they can be searched.

* Use subject_override translation map on marc extract

This changes the horizon index process so that the
MARC stored in solr is ran through our `subject_override`
translation map.

Catalyst displays content from that binary MARC field and
not directly from other solr fields on the record page in
Catalyst.

* Revert "Use subject_override translation map on marc extract"

This reverts commit d5003e485c4e3ae479f323307a6589b8e85658c9.

* Don't translate subject_t it is only used for searching and not viewing

* Index translated subjects into additional fields

* Update subject heading override terms

* Update override yml